Moving the trigger filter response calculation out and simplifying it.
It's generally a stock trapezoidal filter with a risetime and flattop.
It's just scaled by the length to give it a little more scale w/ the
signal you're working with. For that reason, we'll move it to
filters.py with the other general filters. We'll keep the triggering
logic in xia_filters.py since that's not necessarily what people want
to do with a filter.